fix errors

+++ 0x02-python_async_comprehension/0-async_generator.py
@@ -5,7 +5,7 @@ import random
 from typing import Generator
 
 
-async def async_generator() -> Generator[int, None, None]:
+async def async_generator() -> Generator[float, None, None]:
     """Yields random numbers between 1 and 10"""
 
     for i in range(10):
